Commands now check for message delete permissions before attempting to delete other users messages.

+++ src/main/java/net/tempobot/music/commands/CommandHistory.java
@@ -5,6 +5,7 @@ import com.sheepybot.api.entities.command.Arguments;
 import com.sheepybot.api.entities.command.CommandContext;
 import com.sheepybot.api.entities.command.CommandExecutor;
 import com.sheepybot.api.entities.command.parsers.ArgumentParsers;
+import net.dv8tion.jda.api.Permission;
 import net.tempobot.Main;
 import net.tempobot.music.audio.AudioController;
 import net.tempobot.music.audio.TrackScheduler;
@@ -17,7 +18,7 @@ public class CommandHistory implements CommandExecutor {
     public void execute(final CommandContext context,
                         final Arguments args) {
 
-        context.getMessage().delete().queue();
+        if (context.getSelfMember().hasPermission(Permission.MESSAGE_MANAGE)) context.getMessage().delete().queue();
 
         final AudioController controller = Main.get().getAudioLoader().getController(context.getGuild());
         if (controller == null || controller.getPlayer().getPlayingTrack() == null || controller.getPlayer().getPlayingTrack().getState() == AudioTrackState.FINISHED) {
