+++ src/main/java/net/tempobot/music/commands/CommandPlaying.java
@@ -4,6 +4,7 @@ import com.sedmelluq.discord.lavaplayer.track.AudioTrackState;
 import com.sheepybot.api.entities.command.Arguments;
 import com.sheepybot.api.entities.command.CommandContext;
 import com.sheepybot.api.entities.command.CommandExecutor;
+import net.dv8tion.jda.api.Permission;
 import net.tempobot.Main;
 import net.tempobot.music.audio.AudioController;
 
@@ -15,7 +16,7 @@ public class CommandPlaying implements CommandExecutor {
     public void execute(final CommandContext context, 
                         final Arguments args) {
 
-        context.getMessage().delete().queue();
+        if (context.getSelfMember().hasPermission(Permission.MESSAGE_MANAGE)) context.getMessage().delete().queue();
 
         final AudioController controller = Main.get().getAudioLoader().getController(context.getGuild());
         if (controller == null || controller.getPlayer().getPlayingTrack() == null || controller.getPlayer().getPlayingTrack().getState() == AudioTrackState.FINISHED) {
